+++ Changelog.md
@@ -3,6 +3,7 @@
 ## v1.4.0
 
 **Improvements:**
+- Removed accelerometer offset correction from attitude estimators as it must be performed before passing the data
 - Removed ToF attitude correction from `altitudeKF` as it must be performed before passing the data
 
 ## v1.3.2